rename: 디렉토리 이름
- api -> controller 변경

Resolves: #3
