+++ paytogether/src/main/java/com/paytogether/member/service/LoginService.java
@@ -1,6 +1,6 @@
 package com.paytogether.member.service;
 
-import com.paytogether.member.api.response.LoginResponse;
+import com.paytogether.member.controller.response.LoginResponse;
 
 public interface LoginService {
